requirements.txt
asn1crypto==1.5.1
bcrypt==4.0.1
cffi==1.15.1
Click==8.1.3
cryptography==38.0.4
cycler==0.11.0
enum34==1.1.10
Flask==2.2.2
Flask-Login==0.6.2
Flask-WTF==1.0.1
gunicorn==20.1.0
img2pdf==0.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
opencv-contrib-python==4.6.0.66
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
